refactor(controlplane): extract gRPC gateway into gateway package behind Deps interface
Move the agent gRPC stream (Connect / RenewCertificate / ReportEnrollmentSteps
+ receive/dispatch/effects loops, snapshot/runtime handlers, job dispatch) out
of the god-package server into internal/controlplane/gateway (P8.2d, ~1.3k LOC).
The reverse edges — cross-domain ops and everything touching server-only state
(s.mu / revokedAgentIDs / authority / transportSwitchPendingAt /
grpcConnectRateLimiter / enrollmentRec) — go through gateway.Deps, implemented
by *Server in server/gateway_deps.go. server -> gateway; gateway never imports
server. serve.go registers api.Gateway().

Server-state-heavy methods (AuthorizeAgentConnect, ShouldTerminateForRevocation,
MarkTransportSwitchResolved, RenewAgentCertificate, RecordEnrollmentSteps,
HandleInStreamRenewalRequest) moved VERBATIM into gateway_deps.go, preserving
the exact lock discipline (e.g. RenewAgentCertificate calls updateAgentIdentity
under s.mu.Lock()). DTOs agentSnapshot/instanceSnapshot/clientIPSnapshot moved to
gateway (AgentSnapshot/InstanceSnapshot/ClientIPSnapshot); server keeps the two
aliases that still have server-side callers.

server/ non-test LOC: 26510 -> 23116.

No behavior change.
