+++ internal/controlplane/gateway/gateway_messages.go
@@ -1,9 +1,7 @@
-package server
+package gateway
 
 import (
 	"context"
-	"crypto/x509"
-	"encoding/pem"
 	"time"
 
 	"github.com/lost-coder/panvex/internal/controlplane/agenttransport"
@@ -77,25 +75,25 @@ func isPriorityAgentMessage(message *gatewayrpc.ConnectClientMessage) bool {
 	return message.GetJobResult() != nil || message.GetJobAcknowledgement() != nil
 }
 
-func (s *Server) processRegularAgentMessage(
+func (g *Gateway) processRegularAgentMessage(
 	connectionCtx context.Context,
 	agentID string,
 	sess agenttransport.AgentSession,
-	regularSnapshots chan agentSnapshot,
+	regularSnapshots chan AgentSnapshot,
 	message *gatewayrpc.ConnectClientMessage,
 ) error {
 	if hb := message.GetHeartbeat(); hb != nil {
-		s.logger.DebugContext(connectionCtx, logMessageReceived, "agent_id", agentID, "type", "heartbeat")
+		g.logger.DebugContext(connectionCtx, logMessageReceived, "agent_id", agentID, "type", "heartbeat")
 		// P2-LOG-11 / L-11: heartbeat-specific Debug log with the
 		// agent-supplied observed_at so silent-agent troubleshooting can
 		// confirm the panel actually received the heartbeat (vs the
 		// agent never sending it). Debug-level — off by default in
 		// prod, no per-tick spam at Info.
-		s.logger.DebugContext(connectionCtx, "heartbeat received",
+		g.logger.DebugContext(connectionCtx, "heartbeat received",
 			"agent_id", agentID,
 			"observed_at_unix", hb.ObservedAtUnix,
 		)
-		enqueueRegularSnapshot(connectionCtx, regularSnapshots, agentSnapshot{
+		enqueueRegularSnapshot(connectionCtx, regularSnapshots, AgentSnapshot{
 			AgentID:      agentID,
 			NodeName:     hb.NodeName,
 			FleetGroupID: hb.FleetGroupId,
@@ -107,31 +105,31 @@ func (s *Server) processRegularAgentMessage(
 	}
 
 	if snap := message.GetSnapshot(); snap != nil {
-		s.handleSnapshotMessage(connectionCtx, agentID, regularSnapshots, snap)
+		g.handleSnapshotMessage(connectionCtx, agentID, regularSnapshots, snap)
 		return nil
 	}
 
 	if resp := message.GetClientDataResponse(); resp != nil {
-		s.logger.DebugContext(connectionCtx, logMessageReceived, "agent_id", agentID, "type", "client_data_response")
+		g.logger.DebugContext(connectionCtx, logMessageReceived, "agent_id", agentID, "type", "client_data_response")
 		// Run synchronously within the regular message processor goroutine
 		// to prevent unbounded goroutine accumulation from repeated responses.
-		s.reconcileDiscoveredClients(connectionCtx, agentID, resp.GetClients(), resp.GetTelemtUnreachable(), s.now())
+		g.deps.ReconcileDiscoveredClients(connectionCtx, agentID, resp.GetClients(), resp.GetTelemtUnreachable(), g.now())
 		return nil
 	}
 
 	if req := message.GetRenewalRequest(); req != nil {
-		s.logger.DebugContext(connectionCtx, logMessageReceived, "agent_id", agentID, "type", "renewal_request")
-		s.handleInStreamRenewalRequest(connectionCtx, agentID, sess, req)
+		g.logger.DebugContext(connectionCtx, logMessageReceived, "agent_id", agentID, "type", "renewal_request")
+		g.deps.HandleInStreamRenewalRequest(connectionCtx, agentID, sess, req)
 		return nil
 	}
 
 	if batch := message.GetRuntimeEvents(); batch != nil {
-		s.logger.DebugContext(connectionCtx, logMessageReceived, "agent_id", agentID, "type", "runtime_events")
-		s.handleRuntimeEventsBatch(agentID, batch)
+		g.logger.DebugContext(connectionCtx, logMessageReceived, "agent_id", agentID, "type", "runtime_events")
+		g.handleRuntimeEventsBatch(agentID, batch)
 		return nil
 	}
 
-	return s.processPriorityAgentMessage(connectionCtx, agentID, message)
+	return g.processPriorityAgentMessage(connectionCtx, agentID, message)
 }
 
 // handleRuntimeEventsBatch converts a proto RuntimeEventsBatch into the
@@ -143,7 +141,7 @@ func (s *Server) processRegularAgentMessage(
 // nil-check is kept for defence in depth so a future code path that
 // constructs a Server without going through the canonical constructor
 // cannot panic here.
-func (s *Server) handleRuntimeEventsBatch(agentID string, batch *gatewayrpc.RuntimeEventsBatch) {
+func (g *Gateway) handleRuntimeEventsBatch(agentID string, batch *gatewayrpc.RuntimeEventsBatch) {
 	if batch == nil {
 		return
 	}
@@ -174,13 +172,13 @@ func (s *Server) handleRuntimeEventsBatch(agentID string, batch *gatewayrpc.Runt
 	if len(events) == 0 {
 		return
 	}
-	if s.runtimeEvents != nil {
+	if g.runtimeEvents != nil {
 		// Audit #9b: AppendBatch drops reconnect replays via the per-agent
 		// (seq, ts) watermark; publish only what was actually stored so the
 		// dashboard's live feed does not repeat up to 200 events either.
-		events = s.runtimeEvents.AppendBatch(agentID, events)
+		events = g.runtimeEvents.AppendBatch(agentID, events)
 	}
-	if len(events) == 0 || s.events == nil {
+	if len(events) == 0 || g.events == nil {
 		return
 	}
 	// D6a: one bus event per inbound batch instead of one per record. A
@@ -198,7 +196,7 @@ func (s *Server) handleRuntimeEventsBatch(agentID string, batch *gatewayrpc.Runt
 			"fields":   ev.Fields,
 		})
 	}
-	s.events.Publish(eventbus.Event{
+	g.events.Publish(eventbus.Event{
 		Type: cpevents.TypeRuntimeEvents,
 		Data: map[string]any{
 			"agent_id": agentID,
@@ -207,11 +205,11 @@ func (s *Server) handleRuntimeEventsBatch(agentID string, batch *gatewayrpc.Runt
 	})
 }
 
-func (s *Server) processPriorityAgentMessage(ctx context.Context, agentID string, message *gatewayrpc.ConnectClientMessage) error {
-	return s.processPriorityAgentMessageAsync(ctx, nil, nil, agentID, message)
+func (g *Gateway) processPriorityAgentMessage(ctx context.Context, agentID string, message *gatewayrpc.ConnectClientMessage) error {
+	return g.processPriorityAgentMessageAsync(ctx, nil, nil, agentID, message)
 }
 
-func (s *Server) processPriorityAgentMessageAsync(
+func (g *Gateway) processPriorityAgentMessageAsync(
 	connectionCtx context.Context,
 	priorityResultEffects chan<- jobResultEffect,
 	priorityAuditEffects chan<- auditEffect,
@@ -223,9 +221,9 @@ func (s *Server) processPriorityAgentMessageAsync(
 	}
 
 	if jr := message.GetJobResult(); jr != nil {
-		s.logger.DebugContext(connectionCtx, logMessageReceived, "agent_id", agentID, "type", "job_result", "job_id", jr.JobId, "success", jr.Success)
+		g.logger.DebugContext(connectionCtx, logMessageReceived, "agent_id", agentID, "type", "job_result", "job_id", jr.JobId, "success", jr.Success)
 		observedAt := time.Unix(jr.ObservedAtUnix, 0).UTC()
-		s.recordJobResultState(
+		g.recordJobResultState(
 			connectionCtx,
 			agentID,
 			jr.JobId,
@@ -242,7 +240,7 @@ func (s *Server) processPriorityAgentMessageAsync(
 			resultJSON: jr.ResultJson,
 			observedAt: observedAt,
 		}) {
-			s.recordClientJobResultWithContext(connectionCtx, agentID, jr.JobId, jr.Success, jr.Message, jr.ResultJson, observedAt)
+			g.deps.RecordClientJobResult(connectionCtx, agentID, jr.JobId, jr.Success, jr.Message, jr.ResultJson, observedAt)
 		}
 		details := map[string]any{
 			"success": jr.Success,
@@ -254,12 +252,12 @@ func (s *Server) processPriorityAgentMessageAsync(
 			targetID: jr.JobId,
 			details:  details,
 		}) {
-			s.appendAuditWithContext(connectionCtx, agentID, auditJobsResult, jr.JobId, details)
+			g.deps.AppendAudit(connectionCtx, agentID, auditJobsResult, jr.JobId, details)
 		}
 	}
 	if ack := message.GetJobAcknowledgement(); ack != nil {
 		observedAt := time.Unix(ack.ObservedAtUnix, 0).UTC()
-		s.recordJobAcknowledgedState(
+		g.recordJobAcknowledgedState(
 			connectionCtx,
 			agentID,
 			ack.JobId,
@@ -271,113 +269,9 @@ func (s *Server) processPriorityAgentMessageAsync(
 			targetID: ack.JobId,
 			details:  map[string]any{},
 		}) {
-			s.appendAuditWithContext(connectionCtx, agentID, auditJobsAcknowledged, ack.JobId, map[string]any{})
+			g.deps.AppendAudit(connectionCtx, agentID, auditJobsAcknowledged, ack.JobId, map[string]any{})
 		}
 	}
 
 	return nil
-}
-
-// handleInStreamRenewalRequest processes a cert renewal request from an agent
-// over the existing Connect bidi-stream. The response is sent back inline;
-// errors are reported via RenewalResponse.error so the stream stays open.
-func (s *Server) handleInStreamRenewalRequest(ctx context.Context, agentID string, sess agenttransport.AgentSession, req *gatewayrpc.RenewalRequest) {
-	if req.GetAgentId() != agentID {
-		s.logger.WarnContext(ctx, "renewal request agent_id mismatch", "stream_agent_id", agentID, "request_agent_id", req.GetAgentId())
-		_ = sess.Send(&gatewayrpc.ConnectServerMessage{
-			Body: &gatewayrpc.ConnectServerMessage_RenewalResponse{
-				RenewalResponse: &gatewayrpc.RenewalResponse{
-					Error: "agent_id mismatch",
-				},
-			},
-		})
-		return
-	}
-
-	// Reject renewals from revoked agents. Without this, an agent whose
-	// stream is still alive in the window before the 30s revocation watcher
-	// tears it down could obtain a fresh 30-day cert (agentCertificateLifetime) AND have the panel
-	// re-pin its new serial (line ~323) — defeating both revocation and the
-	// serial-pin replay defense. Mirrors the unary RenewCertificate guard.
-	s.mu.RLock()
-	_, revoked := s.revokedAgentIDs[agentID]
-	s.mu.RUnlock()
-	if revoked {
-		s.logger.WarnContext(ctx, "in-stream cert renewal: agent revoked", "agent_id", agentID)
-		_ = sess.Send(&gatewayrpc.ConnectServerMessage{
-			Body: &gatewayrpc.ConnectServerMessage_RenewalResponse{
-				RenewalResponse: &gatewayrpc.RenewalResponse{
-					Error: "agent certificate has been revoked",
-				},
-			},
-		})
-		return
-	}
-
-	csrPEM := req.GetCsrPem()
-	certPEM, caPEM, expiresAt, err := s.authority.SignCSR(csrPEM, agentID, agentCertificateLifetime)
-	if err != nil {
-		s.logger.WarnContext(ctx, "in-stream cert renewal: sign CSR failed", "agent_id", agentID, "error", err)
-		_ = sess.Send(&gatewayrpc.ConnectServerMessage{
-			Body: &gatewayrpc.ConnectServerMessage_RenewalResponse{
-				RenewalResponse: &gatewayrpc.RenewalResponse{
-					Error: err.Error(),
-				},
-			},
-		})
-		return
-	}
-
-	// Parse the new serial so we can update the pin.
-	newSerial := ""
-	if block, _ := pem.Decode([]byte(certPEM)); block != nil {
-		if parsed, parseErr := x509.ParseCertificate(block.Bytes); parseErr == nil {
-			newSerial = parsed.SerialNumber.Text(16)
-		}
-	}
-
-	// Update in-memory cert dates.
-	certIssuedAt := s.now().UTC()
-	certExpiresAtUTC := expiresAt.UTC()
-	s.mu.Lock()
-	if agent, ok := s.updateAgentIdentity(agentID, func(a *Agent) {
-		a.CertIssuedAt = &certIssuedAt
-		a.CertExpiresAt = &certExpiresAtUTC
-		if newSerial != "" {
-			a.CertSerial = newSerial
-		}
-	}); ok && s.batchWriter != nil {
-		s.batchWriter.EnqueueAgent(agentToRecord(agent))
-	}
-	s.mu.Unlock()
-
-	// Persist the new serial so future connects and revocation checks use it.
-	if s.store != nil && newSerial != "" {
-		if err := s.store.UpdateAgentCertSerial(ctx, agentID, newSerial); err != nil {
-			s.logger.WarnContext(ctx, "in-stream cert renewal: persist cert serial failed", "agent_id", agentID, "error", err)
-		}
-	}
-	// In-stream renewal rotates the agent key — must persist the new SPKI pin
-	// or the fail-closed dial verifier (Task 5 / A1) would block the agent
-	// after its first in-stream renewal.
-	s.persistAgentCertPin(ctx, agentID, certPEM)
-
-	sendErr := sess.Send(&gatewayrpc.ConnectServerMessage{
-		Body: &gatewayrpc.ConnectServerMessage_RenewalResponse{
-			RenewalResponse: &gatewayrpc.RenewalResponse{
-				CertificatePem: certPEM,
-				CaPem:          caPEM,
-				ExpiresAtUnix:  expiresAt.Unix(),
-			},
-		},
-	})
-	if sendErr != nil {
-		s.logger.WarnContext(ctx, "in-stream cert renewal: send response failed", "agent_id", agentID, "error", sendErr)
-		return
-	}
-
-	s.logger.InfoContext(ctx, "in-stream cert renewal completed", "agent_id", agentID, "expires_at", expiresAt.UTC())
-	s.appendAuditWithContext(ctx, agentID, auditAgentsCertRenewed, agentID, map[string]any{
-		"expires_at": expiresAt.UTC().Format(time.RFC3339),
-	})
 }
